requirements.txt
colorama==0.4.4
PyGithub==1.55
requests==2.26.0
